backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
invisible-watermark==0.2.0
stegano>=0.11.5
librosa>=0.10.1
soundfile>=0.12.1
scipy>=1.11.0
mutagen>=1.47.0
pypdf>=3.17.0
reportlab>=4.0.0
# httpx removed - no longer needed (NFTs minted in frontend, no backend API calls)

